Move chip to its own component
